+++ src/components/combo_box/combo_box.tsx
@@ -211,7 +211,8 @@ export class EuiComboBox<T> extends Component<
       this.props.selectedOptions,
       initialSearchValue,
       this.props.async,
-      Boolean(this.props.singleSelection)
+      Boolean(this.props.singleSelection),
+      this.props.sortMatchesBy
     ),
     searchValue: initialSearchValue,
     width: 0,
@@ -789,7 +790,12 @@ export class EuiComboBox<T> extends Component<
     nextProps: _EuiComboBoxProps<T>,
     prevState: EuiComboBoxState<T>
   ) {
-    const { options, selectedOptions, singleSelection } = nextProps;
+    const {
+      options,
+      selectedOptions,
+      singleSelection,
+      sortMatchesBy,
+    } = nextProps;
     const { activeOptionIndex, searchValue } = prevState;
 
     // Calculate and cache the options which match the searchValue, because we use this information
@@ -799,7 +805,8 @@ export class EuiComboBox<T> extends Component<
       selectedOptions,
       searchValue,
       nextProps.async,
-      Boolean(singleSelection)
+      Boolean(singleSelection),
+      sortMatchesBy
     );
 
     const stateUpdate: Partial<EuiComboBoxState<T>> = { matchingOptions };
@@ -857,7 +864,12 @@ export class EuiComboBox<T> extends Component<
   };
 
   componentDidUpdate() {
-    const { options, selectedOptions, singleSelection } = this.props;
+    const {
+      options,
+      selectedOptions,
+      singleSelection,
+      sortMatchesBy,
+    } = this.props;
     const { searchValue } = this.state;
 
     // React 16.3 has a bug (fixed in 16.4) where getDerivedStateFromProps
@@ -869,7 +881,8 @@ export class EuiComboBox<T> extends Component<
         selectedOptions,
         searchValue,
         this.props.async,
-        Boolean(singleSelection)
+        Boolean(singleSelection),
+        sortMatchesBy
       )
     );
   }
@@ -918,27 +931,6 @@ export class EuiComboBox<T> extends Component<
       matchingOptions,
     } = this.state;
 
-    let newMatchingOptions = matchingOptions;
-
-    if (sortMatchesBy === 'startsWith') {
-      const refObj: {
-        startWith: Array<EuiComboBoxOptionOption<T>>;
-        others: Array<EuiComboBoxOptionOption<T>>;
-      } = { startWith: [], others: [] };
-
-      newMatchingOptions.forEach(object => {
-        if (
-          object.label
-            .toLowerCase()
-            .startsWith(searchValue.trim().toLowerCase())
-        ) {
-          refObj.startWith.push(object);
-        } else {
-          refObj.others.push(object);
-        }
-      });
-      newMatchingOptions = [...refObj.startWith, ...refObj.others];
-    }
     // Visually indicate the combobox is in an invalid state if it has lost focus but there is text entered in the input.
     // When custom options are disabled and the user leaves the combo box after entering text that does not match any
     // options, this tells the user that they've entered invalid input.
@@ -975,7 +967,7 @@ export class EuiComboBox<T> extends Component<
             fullWidth={fullWidth}
             isLoading={isLoading}
             listRef={this.listRefCallback}
-            matchingOptions={newMatchingOptions}
+            matchingOptions={matchingOptions}
             onCloseList={this.closeList}
             onCreateOption={onCreateOption}
             onOptionClick={this.onOptionClick}